Move Template Content to folder
