+++ src/test/java/uk/gov/hmcts/divorce/document/content/FinalOrderGrantedTemplateContentTest.java
@@ -8,6 +8,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.ccd.sdk.type.YesOrNo;
 import uk.gov.hmcts.divorce.divorcecase.model.ApplicationType;
 import uk.gov.hmcts.divorce.divorcecase.model.DivorceOrDissolution;
+import uk.gov.hmcts.divorce.document.content.templatecontent.FinalOrderGrantedTemplateContent;
 
 import java.time.LocalDateTime;
 import java.util.Map;
@@ -31,20 +32,20 @@ import static uk.gov.hmcts.divorce.document.content.DocmosisTemplateConstants.MA
 import static uk.gov.hmcts.divorce.document.content.DocmosisTemplateConstants.MARRIAGE_DATE;
 import static uk.gov.hmcts.divorce.document.content.DocmosisTemplateConstants.MARRIAGE_OR_CIVIL_PARTNERSHIP;
 import static uk.gov.hmcts.divorce.document.content.DocmosisTemplateConstants.PLACE_OF_MARRIAGE;
-import static uk.gov.hmcts.divorce.document.content.FinalOrderGrantedTemplateContent.A_CIVIL_PARTNERSHIP;
-import static uk.gov.hmcts.divorce.document.content.FinalOrderGrantedTemplateContent.DISSOLUTION_OF_A_CIVIL_PARTNERSHIP;
-import static uk.gov.hmcts.divorce.document.content.FinalOrderGrantedTemplateContent.DISSOLUTION_OF_A_CIVIL_PARTNERSHIP_CY;
-import static uk.gov.hmcts.divorce.document.content.FinalOrderGrantedTemplateContent.DIVORCE;
-import static uk.gov.hmcts.divorce.document.content.FinalOrderGrantedTemplateContent.DIVORCE_CY;
-import static uk.gov.hmcts.divorce.document.content.FinalOrderGrantedTemplateContent.FORMER_CIVIL_PARTNER_CY;
-import static uk.gov.hmcts.divorce.document.content.FinalOrderGrantedTemplateContent.SECTION;
-import static uk.gov.hmcts.divorce.document.content.FinalOrderGrantedTemplateContent.SECTION_18A;
-import static uk.gov.hmcts.divorce.document.content.FinalOrderGrantedTemplateContent.SECTION_18C;
-import static uk.gov.hmcts.divorce.document.content.FinalOrderGrantedTemplateContent.SPOUSE;
-import static uk.gov.hmcts.divorce.document.content.FinalOrderGrantedTemplateContent.SPOUSE_CY;
-import static uk.gov.hmcts.divorce.document.content.FinalOrderGrantedTemplateContent.SPOUSE_OR_CP;
-import static uk.gov.hmcts.divorce.document.content.FinalOrderGrantedTemplateContent.THE_MARRIAGE;
-import static uk.gov.hmcts.divorce.document.content.FinalOrderGrantedTemplateContent.THE_MARRIAGE_OR_CP;
+import static uk.gov.hmcts.divorce.document.content.templatecontent.FinalOrderGrantedTemplateContent.A_CIVIL_PARTNERSHIP;
+import static uk.gov.hmcts.divorce.document.content.templatecontent.FinalOrderGrantedTemplateContent.DISSOLUTION_OF_A_CIVIL_PARTNERSHIP;
+import static uk.gov.hmcts.divorce.document.content.templatecontent.FinalOrderGrantedTemplateContent.DISSOLUTION_OF_A_CIVIL_PARTNERSHIP_CY;
+import static uk.gov.hmcts.divorce.document.content.templatecontent.FinalOrderGrantedTemplateContent.DIVORCE;
+import static uk.gov.hmcts.divorce.document.content.templatecontent.FinalOrderGrantedTemplateContent.DIVORCE_CY;
+import static uk.gov.hmcts.divorce.document.content.templatecontent.FinalOrderGrantedTemplateContent.FORMER_CIVIL_PARTNER_CY;
+import static uk.gov.hmcts.divorce.document.content.templatecontent.FinalOrderGrantedTemplateContent.SECTION;
+import static uk.gov.hmcts.divorce.document.content.templatecontent.FinalOrderGrantedTemplateContent.SECTION_18A;
+import static uk.gov.hmcts.divorce.document.content.templatecontent.FinalOrderGrantedTemplateContent.SECTION_18C;
+import static uk.gov.hmcts.divorce.document.content.templatecontent.FinalOrderGrantedTemplateContent.SPOUSE;
+import static uk.gov.hmcts.divorce.document.content.templatecontent.FinalOrderGrantedTemplateContent.SPOUSE_CY;
+import static uk.gov.hmcts.divorce.document.content.templatecontent.FinalOrderGrantedTemplateContent.SPOUSE_OR_CP;
+import static uk.gov.hmcts.divorce.document.content.templatecontent.FinalOrderGrantedTemplateContent.THE_MARRIAGE;
+import static uk.gov.hmcts.divorce.document.content.templatecontent.FinalOrderGrantedTemplateContent.THE_MARRIAGE_OR_CP;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.FORMATTED_TEST_CASE_ID;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_CASE_ID;
 import static uk.gov.hmcts.divorce.testutil.TestDataHelper.buildCaseDataForGrantFinalOrder;
